[NEW] Rename of the project files
