+++ forecasting/kafka_forecasting/kafka_producer.py
@@ -2,7 +2,7 @@ import json
 import time
 from confluent_kafka import Producer, KafkaError, KafkaException
 
-from forecasting_python.kafka_python.utils import kafka_config
+from forecasting.kafka_forecasting.utils import kafka_config
 
 
 # Check the following Documentation: https://github.com/confluentinc/confluent-kafka-python/blob/master/examples/confluent_cloud.py
